// frontend/src/app/shared/interfaces/user.interface.ts
// Generated by https://quicktype.io

export interface User {
  id: number;
  username: string;
  password: string;
  accessId: string;
  email: string;
  name: string;
  lastName: string;
  role: string;
  accountNonLocked: boolean;
  accountNonExpired: boolean;
  credentialsNonExpired: boolean;
  enabled: boolean;
  failedAttempts: number;
  lockTime: null;
}